More changes -- will update summary
Created using spr 1.3.4
