+++ Cargo.toml
@@ -6,9 +6,9 @@ members = [
     "caboose-util",
     "certificates",
     "common",
-    "crdb-seed",
     "ddm-admin-client",
     "deploy",
+    "dev-tools/crdb-seed",
     "dev-tools/omdb",
     "dev-tools/omicron-dev",
     "dev-tools/xtask",
@@ -76,6 +76,7 @@ default-members = [
     "ddm-admin-client",
     "dpd-client",
     "deploy",
+    "dev-tools/crdb-seed",
     "dev-tools/omdb",
     "dev-tools/omicron-dev",
     "dev-tools/xtask",
