+++ test-utils/Cargo.toml
@@ -12,10 +12,12 @@ camino-tempfile.workspace = true
 dropshot.workspace = true
 futures.workspace = true
 headers.workspace = true
+hex.workspace = true
 http.workspace = true
 libc.workspace = true
 omicron-common.workspace = true
 pem.workspace = true
+ring.workspace = true
 rustls.workspace = true
 slog.workspace = true
 subprocess.workspace = true
