+++ test-utils/src/dev/mod.rs
@@ -9,18 +9,17 @@ pub mod clickhouse;
 pub mod db;
 pub mod dendrite;
 pub mod poll;
+pub mod seed;
 pub mod test_cmds;
 
-use anyhow::ensure;
 use anyhow::{Context, Result};
-use camino::Utf8Path;
 use camino::Utf8PathBuf;
 pub use dropshot::test_util::LogContext;
 use dropshot::ConfigLogging;
 use dropshot::ConfigLoggingIfExists;
 use dropshot::ConfigLoggingLevel;
 use slog::Logger;
-use std::io::{BufReader, BufWriter, Write};
+use std::io::BufReader;
 
 /// Set up a [`dropshot::test_util::LogContext`] appropriate for a test named
 /// `test_name`
@@ -46,84 +45,6 @@ pub enum StorageSource {
     /// Copy the database from a seed tarball, which has previously
     /// been created with `PopulateLatest`.
     CopyFromSeed { input_tar: Utf8PathBuf },
-}
-
-/// Creates a [`db::CockroachInstance`] with a populated storage directory
-/// inside a tarball.
-///
-/// This is intended to optimize subsequent calls to [`test_setup_database`] by
-/// reducing the latency of populating the storage directory.
-pub async fn test_setup_database_seed(
-    log: &Logger,
-    output_tar: &Utf8Path,
-) -> Result<()> {
-    let base_seed_dir = output_tar.parent().unwrap();
-    let tmp_seed_dir = camino_tempfile::Utf8TempDir::new_in(base_seed_dir)
-        .context("failed to create temporary seed directory")?;
-
-    let mut db = setup_database(
-        log,
-        StorageSource::PopulateLatest {
-            output_dir: tmp_seed_dir.path().to_owned(),
-        },
-    )
-    .await
-    .context("failed to setup database")?;
-    db.cleanup().await.context("failed to cleanup database")?;
-
-    // See https://github.com/cockroachdb/cockroach/issues/74231 for context on
-    // this. We use this assertion to check that our seed directory won't point
-    // back to itself, even if it is copied elsewhere.
-    let dirs_record_path = tmp_seed_dir.path().join("temp-dirs-record.txt");
-    let dirs_record_len = dirs_record_path
-        .metadata()
-        .with_context(|| {
-            format!("cannot access metadata for {dirs_record_path}")
-        })?
-        .len();
-    ensure!(
-        dirs_record_len == 0,
-        "Temporary directory record should be empty (was {dirs_record_len}) \
-        after graceful shutdown",
-    );
-
-    // Tar up the directory -- this prevents issues where some but not all of
-    // the files get cleaned up by /tmp cleaners. See
-    // https://github.com/oxidecomputer/omicron/issues/4193.
-    let atomic_file = atomicwrites::AtomicFile::new(
-        &output_tar,
-        // We don't expect this to exist, but if it does, we want to overwrite
-        // it. That is because there's a remote possibility that multiple
-        // instances of test_setup_database_seed are running simultaneously.
-        atomicwrites::OverwriteBehavior::AllowOverwrite,
-    );
-    let res = atomic_file.write(|f| {
-        // Tar up the directory here.
-        let writer = BufWriter::new(f);
-        let mut tar = tar::Builder::new(writer);
-        tar.follow_symlinks(false);
-        tar.append_dir_all(".", tmp_seed_dir.path()).with_context(|| {
-            format!(
-                "failed to append directory `{}` to tarball",
-                tmp_seed_dir.path(),
-            )
-        })?;
-
-        let mut writer =
-            tar.into_inner().context("failed to finish writing tarball")?;
-        writer.flush().context("failed to flush tarball")?;
-
-        Ok::<_, anyhow::Error>(())
-    });
-    match res {
-        Ok(()) => Ok(()),
-        Err(atomicwrites::Error::Internal(error)) => {
-            Err(error).with_context(|| {
-                format!("failed to write seed tarball: `{}`", output_tar)
-            })
-        }
-        Err(atomicwrites::Error::User(error)) => Err(error),
-    }
 }
 
 /// Set up a [`db::CockroachInstance`] for running tests.
